+++ components/external_server/src/routes/report.py
@@ -10,14 +10,13 @@ from pymongo.database import Database
 from shared.database.datamodels import latest_datamodel
 from shared.database.measurements import recent_measurements
 from shared.database.reports import insert_new_report
-from shared.model.report import Report
 from shared.utils.functions import iso_timestamp, report_date_time
 from shared.utils.type import ReportId
 from shared.initialization.secrets import EXPORT_FIELDS_KEYS_NAME
 
 from database.reports import latest_report, latest_reports
 from model.actions import copy_report
-from model.issue_tracker import IssueTracker
+from model.report import Report
 from model.transformations import (
     decrypt_credentials,
     encrypt_credentials,
@@ -214,13 +213,7 @@ def get_report_issue_tracker_suggestions(report_uuid: Report, query: str, databa
     """Get suggestions for issue ids from the issue tracker using the query string."""
     data_model = latest_datamodel(database)
     report = latest_report(database, data_model, report_uuid)
-    issue_tracker_data = report.get("issue_tracker", {})
-    parameters = issue_tracker_data.get("parameters", {})
-    url = parameters.get("url")
-    username = parameters.get("username", "")
-    password = parameters.get("password", "")
-    private_token = parameters.get("private_token", "")
-    issue_tracker = IssueTracker(url, username, password, private_token)
+    issue_tracker = report.issue_tracker()
     return dict(suggestions=[issue.as_dict() for issue in issue_tracker.get_suggestions(query)])