+++ components/external_server/tests/model/test_issue_tracker.py
@@ -11,34 +11,68 @@ class IssueTrackerTest(unittest.TestCase):
     """Unit tests for the issue tracker."""
 
     ISSUE_TRACKER_URL = "https://tracker"
+    PROJECT_KEY = "KEY"
+    ISSUE_TYPE = "BUG"
+    ISSUE_SUMMARY = "Issue summary"
+
+    def setUp(self):
+        """Override to set up the issue tracker."""
+        self.issue_tracker = IssueTracker(self.ISSUE_TRACKER_URL, self.PROJECT_KEY, self.ISSUE_TYPE)
 
     def test_url(self):
         """Test the issue tracker url."""
-        self.assertEqual(self.ISSUE_TRACKER_URL, IssueTracker(self.ISSUE_TRACKER_URL).url)
+        self.assertEqual(self.ISSUE_TRACKER_URL, self.issue_tracker.url)
 
     def test_username_and_password(self):
         """Test the issue tracker credentials."""
-        issue_tracker = IssueTracker(self.ISSUE_TRACKER_URL, "username", "password")
+        issue_tracker = IssueTracker(self.ISSUE_TRACKER_URL, self.PROJECT_KEY, self.ISSUE_TYPE, "username", "password")
         self.assertEqual("username", issue_tracker.username)
         self.assertEqual("password", issue_tracker.password)
 
     def test_private_token(self):
         """Test the issue tracker credentials."""
-        issue_tracker = IssueTracker(self.ISSUE_TRACKER_URL, private_token="token")
+        issue_tracker = IssueTracker(self.ISSUE_TRACKER_URL, self.PROJECT_KEY, self.ISSUE_TYPE, private_token="token")
         self.assertEqual("token", issue_tracker.private_token)
 
     @patch("requests.get")
     def test_get_suggestions(self, requests_get):
         """Test that issue suggestions are returned."""
         response = Mock()
-        response.json.return_value = dict(issues=[dict(key="FOO-42", fields=dict(summary="Summary"))])
+        response.json.return_value = dict(issues=[dict(key="FOO-42", fields=dict(summary=self.ISSUE_SUMMARY))])
         requests_get.return_value = response
-        issue_tracker = IssueTracker(self.ISSUE_TRACKER_URL)
-        self.assertEqual([IssueSuggestion("FOO-42", "Summary")], issue_tracker.get_suggestions("Summ"))
+        self.assertEqual([IssueSuggestion("FOO-42", "Issue summary")], self.issue_tracker.get_suggestions("Summ"))
 
     def test_get_suggestions_without_url(self):
         """Test that an empty list of issue suggestions is returned."""
-        issue_tracker = IssueTracker("")
         logging.disable(logging.CRITICAL)
-        self.assertEqual([], issue_tracker.get_suggestions("Query"))
+        self.assertEqual([], self.issue_tracker.get_suggestions("Query"))
+        logging.disable(logging.NOTSET)
+
+    @patch("requests.post")
+    def test_create_issue(self, requests_post):
+        """Test that an issue can be created."""
+        response = Mock()
+        response.json.return_value = dict(key="FOO-42")
+        requests_post.return_value = response
+        self.assertEqual(("FOO-42", ""), self.issue_tracker.create_issue(self.ISSUE_SUMMARY))
+
+    def test_create_issue_with_invalid_url(self):
+        """Test that without a valid URL an error message is returned."""
+        issue_tracker = IssueTracker("invalid", self.PROJECT_KEY, self.ISSUE_TYPE)
+        logging.disable(logging.CRITICAL)
+        self.assertEqual(
+            (
+                "",
+                "Invalid URL 'invalid/rest/api/2/issue': No scheme supplied. Perhaps you meant "
+                "http://invalid/rest/api/2/issue?",
+            ),
+            issue_tracker.create_issue("New issue"),
+        )
+        logging.disable(logging.NOTSET)
+
+    def test_create_issue_withoutd_url(self):
+        """Test that without a URL an error message is returned."""
+        issue_tracker = IssueTracker("", "", "")
+        logging.disable(logging.CRITICAL)
+        self.assertEqual(("", "Issue tracker has no URL configured."), issue_tracker.create_issue(self.ISSUE_SUMMARY))
         logging.disable(logging.NOTSET)